fixed torch integer division problem

+++ Tools/analyzeLatentVector_LogisticRegression.py
@@ -128,7 +128,7 @@ def main():
 
         sigmoidX = 1.0 / (1.0 + np.exp(-(W0Ex + W1Ex * X)))
         predictX = (sigmoidX >= 0.5).astype(np.int)
-        Y = np.reshape(np.repeat(Y01, F * H * W, axis=0), X.shape)
+        Y = np.reshape(np.repeat(Y01, F * H * W, axis=0), X.shape).astype(np.int)
         accuracyX = ((predictX - Y) == 0).sum(axis=3)*1.0 / N
 
         minAccuracy.append(accuracyX.min())
